% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GART_fun.R
\name{GART}
\alias{GART}
\title{Run GART main function}
\usage{
GART(
  data,
  base_names = "weighted",
  tau = NULL,
  is_benchmark = F,
  is_valid = F,
  data_valid = NULL
)
}
\arguments{
\item{data}{A list including two elements named `target` and `source`. Both contain two sublists named `X` and `y`, corresponding to the training data of source sites and the target site. Note that `X` and `y` in `source` should include the same number of sublists if multiple source sites exist.}

\item{base_names}{A vector of names representing different initial estimators for GART. Possible choices include 'target' (target only estimator), 'zero', 'convex' (convex combination of sources and target), 'weighted' (weighted average of sources and target), 'source' (convex combination of sources).}

\item{tau}{A number constroling the size of constraint set in GART. Default is 1/n_target.}

\item{is_benchmark}{If \code{TRUE}, calculate benchmark methods including target only estimator, source mixture, maximin, transLasso and transGLM. Default is \code{FALSE}.}

\item{is_valid}{If \code{TRUE}, evaluate model performance based on `data_valid`. Default is \code{FALSE}.}

\item{data_valid}{A list including three elements for the validation data if the true coefficient generating validation data is known, namely `X` (an N by p matrix), `y` (a vector of length N) and `beta` (a vector of length p or p+1 (with intercept)). If the true coefficient is unknown, only the first two elements are enough.}
}
\value{
A list containing estimation and/or evaluation results by running GART (and other benchmark methods if `is_benchmark=TRUE`).
}
\description{
Run GART main function
}
