% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{simu_data}
\alias{simu_data}
\title{Generate simulation data to run GART}
\usage{
simu_data(alpha = 0.07)
}
\arguments{
\item{alpha}{A numeric parameter controling the distance between the source convex hull and the target model.}
}
\value{
A list of two elements including the input data for GART estimation and a validation dataset for GART evaluation. Note that for illustration purpose, the validation population shares the same linear model as the training target population.
}
\description{
Generate simulation data to run GART
}
